Moved bot.token to application.properties, created class BotConfig
