+++ src/main/java/com/example/demo_hh_stat/Bot.java
@@ -1,11 +1,14 @@
 package com.example.demo_hh_stat;
 
 
-import com.example.demo_hh_stat.entity.Filter;
+import com.example.demo_hh_stat.entity.Vacancy;
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.UpdatesListener;
 import com.pengrad.telegrambot.request.SendMessage;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+
+import java.util.List;
 
 
 @Component
@@ -14,10 +17,11 @@ public class Bot {
     private final Logic logic;
 
     //TODO убрать токен в config и вытягивать его в классе-конфигураторе
-    private String botToken = "";
+    private String botToken;
 
-    public Bot(Logic logic) {
+    public Bot(Logic logic, @Value("${bot.token}") String botToken) {
         this.logic = logic;
+        this.botToken = botToken;
 
         TelegramBot bot = new TelegramBot(botToken);
         bot.setUpdatesListener(element -> {
@@ -39,10 +43,19 @@ public class Bot {
                             .region(massive[1])
                             .requestId(requestId)
                             .build();
-                    int numVacancies = logic.getNumberOfVacancies(filter);
-                    int allResponses = logic.getAllResponses(filter);
-                    bot.execute(new SendMessage(it.message().chat().id(),
-                            "Найдено вакансий " + numVacancies + " и всего откликов: " + allResponses));
+
+                    List<Vacancy> list = logic.getVacancyFilter(filter);
+                    int allResponses = getAllResponses(list);
+                    if (list.size() != 0){
+                        //show all vacancies
+                        /*list.forEach(vacancy -> {
+                            bot.execute(new SendMessage(it.message().chat().id(), "Вакансия: " + vacancy.getName() + "\nКоличество откликов: " + vacancy.getCounters().getResponses() + "\nСсылка: http://hh.ru/vacancy/" + vacancy.getId()));
+                            System.out.println(vacancy.getId() + " " + vacancy.getName());
+                        });*/
+                        bot.execute(new SendMessage(it.message().chat().id(), "Количество найденных вакансий: " + list.size() + " и всего откликов " + allResponses + "."));
+                    } else {
+                        bot.execute(new SendMessage(it.message().chat().id(), "По вашему запросу вакансий не найдено."));
+                    }
                 }
             });
             return UpdatesListener.CONFIRMED_UPDATES_ALL;
@@ -55,5 +68,13 @@ public class Bot {
             reg = reg + " " + s[i];
         }
         return reg;
+    }
+
+    public int getAllResponses(List<Vacancy> list){
+        int totalResponses = 0;
+        for (Vacancy v : list) {
+            totalResponses += v.getCounters().getResponses();
+        }
+        return totalResponses;
     }
 }
